+++ CHANGELOG.md
@@ -16,10 +16,11 @@ Changes in this release:
 - Added support for /etc/inmanta.d style configuration files (#183)
 - Increased the iteration limit to 10000. This value is controlled with INMANTA_MAX_ITERATIONS
   environment variable.
-- Added support for custom resource deserialization by adding the 'populate' method 
+- Added support for custom resource deserialization by adding the 'populate' method
 - Improve compiler scaling by using more efficient data structures
 - Added the --export-plugin option to the export command (#1277)
 - Only one of set_created, set_updated or set_purged may be called now from a handler
+- Remove facts when the resource is no longer present in any version (#1027)
 
 DEPRECATIONS:
 * The files /etc/inmanta/agent.cfg and /etc/inmanta/server.cfg are not used anymore. More information about the available